Rename form files + memoise api fetch
